% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jade.R
\name{renderer}
\alias{renderer}
\title{Renderer}
\usage{
renderer(file, data)
}
\arguments{
\item{file}{File to render.}

\item{data}{Data to render variables.}
}
\description{
Renders file using specified data.
}
